+++ changelog.md
@@ -38,6 +38,7 @@ This version is not yet released. If you are reading this on the website, then t
 - Add a `Lorem` constant, which contains the Lorem Ipsum text
 - Add an experimental [`sort ⍆`](https://uiua.org/docs/sort) function
   - Sorting is very common
+- Add the experimental [`layout`](https://uiua.org/docs/layout) function, which renders text into an image array
 - Remove previously deprecated function strands
 ### Interpreter
 - Add the `uiua find` command, which finds Uiua code that matches the given unformatted text